requirements.txt
Django==4.2.4
WebOb==1.8.7
annotated-types==0.5.0
asgiref==3.7.2
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-static==4.2.1
django-bootstrap5==23.3
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-scheduler==0.10.1
django-selenium==0.9.8
django-static-fontawesome==6.4.0.0
django-static-jquery==2.1.4
djangoajax==3.3
js.jquery==3.6.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
pydantic==2.2.1
pydantic-core==2.6.1
python-dateutil==2.8.1
pytz==2023.3
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3